Split radio and simplify slots
