+++ vue/components/ui/atoms/checkbox/checkbox.stories.js
@@ -5,9 +5,6 @@ storiesOf("Atoms", module)
     .addDecorator(withKnobs)
     .add("Checkbox", () => ({
         props: {
-            error: {
-                default: boolean("Error", false)
-            },
             disabled: {
                 default: boolean("Disabled", false)
             },
@@ -16,9 +13,6 @@ storiesOf("Atoms", module)
             },
             label: {
                 default: text("Label", "Checkbox")
-            },
-            value: {
-                default: text("Value", "checkbox")
             },
             icon: {
                 default: select(
@@ -30,8 +24,15 @@ storiesOf("Atoms", module)
                     "check"
                 )
             },
-            slots: {
-                default: boolean("Slots", false)
+            variant: {
+                default: select(
+                    "Variant",
+                    {
+                        Unset: null,
+                        Error: "error"
+                    },
+                    null
+                )
             },
             size: {
                 default: number("Size", 4)
@@ -51,20 +52,12 @@ storiesOf("Atoms", module)
             <div>
                 <checkbox
                     v-bind:label="label"
-                    v-bind:value="value"
                     v-bind:checked.sync="checkedData"
                     v-bind:disabled="disabled"
-                    v-bind:error="error"
                     v-bind:icon="icon"
                     v-bind:size="size"
-                >
-                    <template v-slot:before-item="{ label, value, checked }" v-if="slots">
-                        <p>Custom before checkbox {{ label }} with value {{ value }} and checked is {{ checked }}</p>
-                    </template>
-                    <template v-slot:after-item="{ label, value, checked }" v-if="slots">
-                        <p>Custom After checkbox {{ label }}</p>
-                    </template>
-                </checkbox>
+                    v-bind:variant="variant"
+                />
                 <p>Checked: {{ checkedData }}</p>
             </div>
         `
