+++ vue/components/ui/atoms/radio/radio.vue
@@ -1,39 +1,27 @@
 <template>
-    <div class="radio">
-        <div
-            class="choice"
-            v-bind:class="{
-                disabled: disabled || item.disabled,
-                error: error || item.error,
-                checked: item.value === value
-            }"
-            v-bind:index="index"
-            v-bind:tabindex="_getTabIndex(item, index)"
-            v-for="(item, index) in items"
-            v-bind:key="item.value"
-            v-bind:ref="`choice-${index}`"
-            v-on:keydown.space="onSpacebar(item)"
-            v-on:keydown.up="onArrowUp(index)"
-            v-on:keydown.down="onArrowDown(index)"
-            v-on:click="onClick(item)"
-        >
-            <slot name="before-item" v-bind:item="item" v-bind:index="index" />
-            <div class="radio-input">
-                <input type="radio" class="value" v-bind:id="item.value" />
-                <div class="radio-circle" />
-                <label v-bind:for="item.value" class="label">
-                    {{ item.label ? item.label : item.value }}
-                </label>
-            </div>
-            <slot name="after-item" v-bind:item="item" v-bind:index="index" />
+    <div
+        class="radio"
+        v-bind:class="classes"
+        ref="root"
+        v-on:click="onClick"
+        v-on:keydown="onKeydown"
+    >
+        <slot name="before-item" />
+        <div class="radio-input">
+            <input type="radio" class="value" v-bind:id="value" />
+            <div class="radio-circle" />
+            <label v-bind:for="value" class="label" v-if="label">
+                {{ label }}
+            </label>
         </div>
+        <slot name="after-item" />
     </div>
 </template>
 
 <style lang="scss" scoped>
 @import "css/variables.scss";
 
-.choice {
+.radio {
     display: block;
     line-height: 13px;
     outline: none;
@@ -42,15 +30,15 @@
     width: fit-content;
 }
 
-.choice > .radio-input {
+.radio > .radio-input {
     display: inline-block;
 }
 
-.choice > .radio-input > .value {
+.radio > .radio-input > .value {
     display: none;
 }
 
-.choice > .radio-input > .radio-circle {
+.radio > .radio-input > .radio-circle {
     background-color: #fafbfc;
     border: 2px solid #dfe1e5;
     border-radius: 50% 50% 50% 50%;
@@ -62,45 +50,45 @@
     width: 4px;
 }
 
-.choice:not(.disabled):not(.error):active > .radio-input > .radio-circle {
+.radio:not(.disabled):not(.error):active > .radio-input > .radio-circle {
     background: url("~./assets/check-dark.svg") center / 4px no-repeat #f4f5f7;
     border: 2px solid #c3c9cf;
     padding: 3px 3px 3px 3px;
 }
 
-.choice.error > .radio-input > .radio-circle {
+.radio.error > .radio-input > .radio-circle {
     background-color: #f4f5f7;
     border: 2px solid $dark-red;
 }
 
-.choice.disabled > .radio-input > .radio-circle {
+.radio.disabled > .radio-input > .radio-circle {
     background: #f4f5f7;
     border: 2px solid #f4f5f7;
     cursor: default;
 }
 
-.choice.checked > .radio-input > .radio-circle {
+.radio.checked > .radio-input > .radio-circle {
     background: url("~./assets/check.svg") center / 4px no-repeat $dark;
     border: 2px solid $dark;
     padding: 3px 3px 3px 3px;
 }
 
-.choice.error.checked > .radio-input > .radio-circle {
+.radio.error.checked > .radio-input > .radio-circle {
     background: url("~./assets/check.svg") center / 4px no-repeat $dark;
     border: 2px solid $dark-red;
 }
 
-.choice.disabled.checked > .radio-input > .radio-circle {
+.radio.disabled.checked > .radio-input > .radio-circle {
     background: url("~./assets/check-gray.svg") center / 4px no-repeat #f4f5f7;
     border: 2px solid #f6f7f9;
     padding: 3px 3px 3px 3px;
 }
 
-.choice:focus:not(.disabled) > .radio-input > .radio-circle {
+.radio:focus:not(.disabled) > .radio-input > .radio-circle {
     border-color: $aqcua-blue;
 }
 
-.choice > .radio-input > .label {
+.radio > .radio-input > .label {
     color: $grey;
     cursor: pointer;
     display: inline-block;
@@ -110,7 +98,7 @@
     vertical-align: middle;
 }
 
-.choice.disabled > .radio-input > .label {
+.radio.disabled > .radio-input > .label {
     cursor: default;
 }
 </style>
@@ -119,73 +107,45 @@
 export const Radio = {
     name: "radio",
     props: {
-        items: {
-            type: Array,
-            default: () => []
+        label: {
+            type: String,
+            default: null
         },
         value: {
             type: String,
             default: null
         },
+        checked: {
+            type: Boolean,
+            default: false
+        },
         disabled: {
             type: Boolean,
             default: false
         },
-        error: {
-            type: Boolean,
-            default: false
+        variant: {
+            type: String,
+            default: null
         }
     },
     computed: {
-        firstEnabledIndex() {
-            if (this.disabled) return null;
+        classes() {
+            const base = {
+                disabled: this.disabled,
+                checked: this.checked
+            };
 
-            const firstEnabled = this.items
-                .map((item, index) => [item, index])
-                .find(([item, index]) => !item.disabled);
+            if (this.variant) base[this.variant] = true;
 
-            return firstEnabled ? firstEnabled[1] : null;
+            return base;
         }
     },
     methods: {
-        focusAndSelectItem(item, index) {
-            if (item.disabled || this.disabled) return false;
-            if (index !== undefined) this.$refs[`choice-${index}`][0].focus();
-            this.selectItem(item);
-            return true;
+        onClick(event) {
+            this.$emit("click", event);
         },
-        selectItem(item) {
-            if (item.disabled || this.disabled) return false;
-            if (this.value === item.value) return false;
-            this.$emit("update:value", item.value);
-            return true;
-        },
-        onClick(item) {
-            this.selectItem(item);
-        },
-        onSpacebar(item) {
-            this.selectItem(item);
-        },
-        onArrowUp(index) {
-            for (let i = index - 1; i > index - this.items.length; i--) {
-                const _index = this._negativeModulo(i, this.items.length);
-                const item = this.items[_index];
-                if (this.focusAndSelectItem(item, _index)) return;
-            }
-        },
-        onArrowDown(index) {
-            for (let i = index + 1; i < index + this.items.length; i++) {
-                const _index = this._negativeModulo(i, this.items.length);
-                const item = this.items[_index];
-                if (this.focusAndSelectItem(item, _index)) return;
-            }
-        },
-        _getTabIndex(item, index) {
-            if (this.disabled || item.disabled) return null;
-            return this.firstEnabledIndex === index ? 0 : -1;
-        },
-        _negativeModulo(number, modulo) {
-            return ((number % modulo) + modulo) % modulo;
+        onKeydown(event) {
+            this.$emit("keydown", event);
         }
     }
 };
