+++ vue/components/ui/molecules/index.js
@@ -12,6 +12,7 @@ import { Lightbox } from "./lightbox/lightbox.vue";
 import { Lineup } from "./lineup/lineup.vue";
 import { Modal } from "./modal/modal.vue";
 import { ProgressBar } from "./progress-bar/progress-bar.vue";
+import { RadioGroup } from "./radio-group/radio-group.vue";
 import { Table } from "./table/table.vue";
 import { Tabs } from "./tabs/tabs.vue";
 import { Search } from "./search/search.vue";
@@ -33,6 +34,7 @@ const install = Vue => {
     Vue.component("lineup", Lineup);
     Vue.component("modal", Modal);
     Vue.component("progress-bar", ProgressBar);
+    Vue.component("radio-group", RadioGroup);
     Vue.component("search", Search);
     Vue.component("select-ripe", Select);
     Vue.component("table-ripe", Table);
@@ -55,6 +57,7 @@ export {
     Lineup,
     Modal,
     ProgressBar,
+    RadioGroup,
     Search,
     Select,
     Table,
